+++ test.mjs
@@ -9,7 +9,6 @@ import test_cli_coverage_node from "./coverage-node.test.mjs";
 import test_reportCliError from "./reportCliError.test.mjs";
 import test_semver from "./semver.test.mjs";
 import test_sourceRange from "./sourceRange.test.mjs";
-import test_replaceStackTraces from "./test/replaceStackTraces.test.mjs";
 
 const tests = new TestDirector();
 
@@ -20,6 +19,5 @@ test_cli_coverage_node(tests);
 test_reportCliError(tests);
 test_semver(tests);
 test_sourceRange(tests);
-test_replaceStackTraces(tests);
 
 tests.run();